+++ app/Werewolf/Commands/Quit.hs
@@ -24,7 +24,7 @@ import Control.Monad.Writer
 import Data.Text (Text)
 
 import Game.Werewolf.Command
-import Game.Werewolf.Engine   hiding (isVillagersTurn)
+import Game.Werewolf.Engine
 import Game.Werewolf.Response
 
 -- | Handle.
@@ -38,6 +38,6 @@ handle callerName = do
 
     let command = quitCommand callerName
 
-    case runExcept (runWriterT $ execStateT (apply command >> checkTurn >> checkGameOver) game) of
+    case runExcept (runWriterT $ execStateT (apply command >> checkStage >> checkGameOver) game) of
         Left errorMessages      -> exitWith failure { messages = errorMessages }
         Right (game', messages) -> writeGame game' >> exitWith success { messages = messages }