+++ app/Werewolf/Commands/Vote.hs
@@ -50,6 +50,6 @@ handle callerName (Options targetName) = do
             else lynchVoteCommand
             ) callerName targetName
 
-    case runExcept (runWriterT $ execStateT (apply command >> checkTurn >> checkGameOver) game) of
+    case runExcept (runWriterT $ execStateT (apply command >> checkStage >> checkGameOver) game) of
         Left errorMessages      -> exitWith failure { messages = errorMessages }
         Right (game', messages) -> writeGame game' >> exitWith success { messages = messages }