+++ src/Game/Werewolf/Response.hs
@@ -28,36 +28,35 @@ module Game.Werewolf.Response (
     publicMessage, privateMessage,
 
     -- ** Generic messages
-    newGameMessages, turnMessages, nightFallsMessage, gameOverMessages, playerQuitMessage,
+    newGameMessages, stageMessages, gameOverMessages, playerQuitMessage,
 
     -- ** Ping messages
     pingSeersMessage, pingWerewolvesMessage,
 
     -- ** Status messages
-    currentTurnMessage, rolesInGameMessage, playersInGameMessage, waitingOnMessage,
+    currentStageMessages, rolesInGameMessage, playersInGameMessage, waitingOnMessage,
 
     -- ** Seers turn messages
-    seersTurnMessages, playerSeenMessage,
+    playerSeenMessage,
 
-    -- ** Villagers turn messages
-    villagersTurnMessage, playerMadeLynchVoteMessage, playerLynchedMessage, noPlayerLynchedMessage,
+    -- ** Villages' turn messages
+    playerMadeLynchVoteMessage, playerLynchedMessage, noPlayerLynchedMessage,
     scapegoatLynchedMessage,
 
-    -- ** Werewolves turn messages
-    werewolvesTurnMessages, playerMadeDevourVoteMessage, playerDevouredMessage,
-    noPlayerDevouredMessage, villagersLynchVoteMessage,
+    -- ** Werewolves' turn messages
+    playerMadeDevourVoteMessage, playerDevouredMessage, noPlayerDevouredMessage,
 
     -- ** Generic error messages
     gameIsOverMessage, playerDoesNotExistMessage, playerCannotDoThatMessage,
     playerCannotDoThatRightNowMessage, playerIsDeadMessage, roleDoesNotExistMessage,
 
-    -- ** Seers turn error messages
+    -- ** Seers' turn error messages
     playerHasAlreadySeenMessage,
 
     -- ** Voting turn error messages
     playerHasAlreadyVotedMessage, targetIsDeadMessage,
 
-    -- ** Werewolves turn error messages
+    -- ** Werewolves' turn error messages
     playerCannotDevourAnotherWerewolf,
 ) where
 
@@ -130,9 +129,9 @@ privateMessage :: [Text] -> Text -> Message
 privateMessage to = Message (Just to)
 
 newGameMessages :: Game -> [Message]
-newGameMessages game = [newPlayersInGameMessage players', rolesInGameMessage Nothing $ map _role players'] ++ map (newPlayerMessage players') players' ++ turnMessages turn' players'
+newGameMessages game = [newPlayersInGameMessage players', rolesInGameMessage Nothing $ map _role players'] ++ map (newPlayerMessage players') players' ++ stageMessages stage' players'
     where
-        turn'       = game ^. turn
+        stage'      = game ^. stage
         players'    = game ^. players
 
 newPlayersInGameMessage :: [Player] -> Message
@@ -150,16 +149,34 @@ newPlayerMessage players player
             | length (filterWerewolves players) <= 1    = "."
             | otherwise                                 = T.concat [", along with ", T.intercalate ", " (map _name $ filterWerewolves players \\ [player]), "."]
 
-turnMessages :: Turn -> [Player] -> [Message]
-turnMessages NightFalling _     = [nightFallsMessage]
-turnMessages DayBreaking _      = []
-turnMessages Seers players      = seersTurnMessages $ filter isSeer players
-turnMessages Villagers _        = [villagersTurnMessage]
-turnMessages Werewolves players = werewolvesTurnMessages $ filter isWerewolf players
-turnMessages NoOne _            = []
+stageMessages :: Stage -> [Player] -> [Message]
+stageMessages GameOver _                    = []
+stageMessages SeersTurn alivePlayers        = seersTurnMessages $ filterSeers alivePlayers
+stageMessages Sunrise _                     = [sunriseMessage]
+stageMessages Sunset _                      = [nightFallsMessage]
+stageMessages VillagesTurn _                = [villagesTurnMessage]
+stageMessages WerewolvesTurn alivePlayers   = werewolvesTurnMessages $ filterWerewolves alivePlayers
+
+seersTurnMessages :: [Player] -> [Message]
+seersTurnMessages seers = [
+    publicMessage "The Seers wake up.",
+    privateMessage (map _name seers) "Whose allegiance would you like to see?"
+    ]
+
+sunriseMessage :: Message
+sunriseMessage = publicMessage "The sun rises. Everybody wakes up and opens their eyes..."
 
 nightFallsMessage :: Message
 nightFallsMessage = publicMessage "Night falls, the townsfolk are asleep."
+
+villagesTurnMessage :: Message
+villagesTurnMessage = publicMessage "Whom would you like to lynch?"
+
+werewolvesTurnMessages :: [Player] -> [Message]
+werewolvesTurnMessages werewolves = [
+    publicMessage "The Werewolves wake up, recognise one another and choose a new victim.",
+    privateMessage (map _name werewolves) "Whom would you like to devour?"
+    ]
 
 gameOverMessages :: Game -> [Message]
 gameOverMessages game = case aliveAllegiances of
@@ -188,11 +205,13 @@ pingSeersMessage = publicMessage "Waiting on the Seers..."
 pingWerewolvesMessage :: Message
 pingWerewolvesMessage = publicMessage "Waiting on the Werewolves..."
 
-currentTurnMessage :: Text -> Turn -> Message
-currentTurnMessage name NoOne   = gameIsOverMessage name
-currentTurnMessage name turn    = privateMessage [name] $ T.concat [
+currentStageMessages :: Text -> Stage -> [Message]
+currentStageMessages name GameOver  = [gameIsOverMessage name]
+currentStageMessages _ Sunrise      = []
+currentStageMessages _ Sunset       = []
+currentStageMessages name turn      = [privateMessage [name] $ T.concat [
     "It's currently the ", T.pack $ show turn, "' turn."
-    ]
+    ]]
 
 rolesInGameMessage :: Maybe [Text] -> [Role] -> Message
 rolesInGameMessage mTo roles = Message mTo $ T.concat [
@@ -226,20 +245,8 @@ waitingOnMessage to players = Message to $ T.concat [
     where
         playerNames = map _name players
 
-seersTurnMessages :: [Player] -> [Message]
-seersTurnMessages seers = [
-    publicMessage "The Seers wake up.",
-    privateMessage (map _name seers) "Whose allegiance would you like to see?"
-    ]
-
 playerSeenMessage :: Text -> Player -> Message
 playerSeenMessage seerName target = privateMessage [seerName] $ T.concat [target ^. name, " is aligned with the ", T.pack . show $ target ^. role . allegiance, "."]
-
-villagersTurnMessage :: Message
-villagersTurnMessage = publicMessage "The sun rises. Everybody wakes up and opens their eyes..."
-
-villagersLynchVoteMessage :: Message
-villagersLynchVoteMessage = publicMessage "Whom would you like to lynch?"
 
 playerMadeLynchVoteMessage :: Text -> Text -> Message
 playerMadeLynchVoteMessage voterName targetName = publicMessage $ T.concat [voterName, " voted to lynch ", targetName, "."]
@@ -258,12 +265,6 @@ playerLynchedMessage name roleName      = publicMessage $ T.concat [
 
 noPlayerLynchedMessage :: Message
 noPlayerLynchedMessage = publicMessage "Daylight is wasted as the townsfolk squabble over whom to tie up. Looks like no one is being burned this day."
-
-werewolvesTurnMessages :: [Player] -> [Message]
-werewolvesTurnMessages werewolves = [
-    publicMessage "The Werewolves wake up, recognise one another and choose a new victim.",
-    privateMessage (map _name werewolves) "Whom would you like to devour?"
-    ]
 
 scapegoatLynchedMessage :: Text -> Message
 scapegoatLynchedMessage name = publicMessage $ T.unwords [
@@ -285,7 +286,8 @@ playerDevouredMessage name roleName = publicMessage $ T.concat [
 
 noPlayerDevouredMessage :: Message
 noPlayerDevouredMessage = publicMessage $ T.unwords [
-    "Surprisingly you see everyone present at the town square. Perhaps the Werewolves have left Miller's Hollow?",
+    "Surprisingly you see everyone present at the town square.",
+    "Perhaps the Werewolves have left Miller's Hollow?",
     "Still got to keep up the tradition though, so the town clerk calls for a vote."
     ]