+++ test/app/Main.hs
@@ -44,7 +44,7 @@ allCommandTests = [
     testProperty "PROP: lynch vote command errors when target does not exist" prop_lynchVoteCommandErrorsWhenTargetDoesNotExist,
     testProperty "PROP: lynch vote command errors when caller is dead" prop_lynchVoteCommandErrorsWhenCallerIsDead,
     testProperty "PROP: lynch vote command errors when target is dead" prop_lynchVoteCommandErrorsWhenTargetIsDead,
-    testProperty "PROP: lynch vote command errors when not villagers turn" prop_lynchVoteCommandErrorsWhenNotVillagersTurn,
+    testProperty "PROP: lynch vote command errors when not villages turn" prop_lynchVoteCommandErrorsWhenNotVillagesTurn,
     testProperty "PROP: lynch vote command errors when caller has voted" prop_lynchVoteCommandErrorsWhenCallerHasVoted,
     testProperty "PROP: lynch vote command updates votes" prop_lynchVoteCommandUpdatesVotes,
 
@@ -69,30 +69,30 @@ allCommandTests = [
 
 allEngineTests :: [TestTree]
 allEngineTests = [
-    testProperty "PROP: check turn skips seers when no seers" prop_checkTurnSkipsSeersWhenNoSeers,
-    testProperty "PROP: check turn does nothing when game over" prop_checkTurnDoesNothingWhenGameOver,
+    testProperty "PROP: check stage skips seers when no seers" prop_checkStageSkipsSeersWhenNoSeers,
+    testProperty "PROP: check stage does nothing when game over" prop_checkStageDoesNothingWhenGameOver,
 
     testProperty "PROP: check seers turn advances to werewolves" prop_checkSeersTurnAdvancesToWerewolves,
     testProperty "PROP: check seers turn resets sees" prop_checkSeersTurnResetsSees,
     testProperty "PROP: check seers turn does nothing unless all seen" prop_checkSeersTurnDoesNothingUnlessAllSeen,
 
-    testProperty "PROP: check villagers turn advances to seers" prop_checkVillagersTurnAdvancesToSeers,
-    testProperty "PROP: check villagers turn lynches one player when consensus" prop_checkVillagersTurnLynchesOnePlayerWhenConsensus,
-    testProperty "PROP: check villagers turn lynches no one when conflicted and no scapegoats" prop_checkVillagersTurnLynchesNoOneWhenConflictedAndNoScapegoats,
-    testProperty "PROP: check villagers turn lynches scapegoat when conflicted" prop_checkVillagersTurnLynchesScapegoatWhenConflicted,
-    testProperty "PROP: check villagers turn resets votes" prop_checkVillagersTurnResetsVotes,
-    testProperty "PROP: check villagers turn does nothing unless all voted" prop_checkVillagersTurnDoesNothingUnlessAllVoted,
+    testProperty "PROP: check villages turn advances to seers" prop_checkVillagesTurnAdvancesToSeers,
+    testProperty "PROP: check villages turn lynches one player when consensus" prop_checkVillagesTurnLynchesOnePlayerWhenConsensus,
+    testProperty "PROP: check villages turn lynches no one when conflicted and no scapegoats" prop_checkVillagesTurnLynchesNoOneWhenConflictedAndNoScapegoats,
+    testProperty "PROP: check villages turn lynches scapegoat when conflicted" prop_checkVillagesTurnLynchesScapegoatWhenConflicted,
+    testProperty "PROP: check villages turn resets votes" prop_checkVillagesTurnResetsVotes,
+    testProperty "PROP: check villages turn does nothing unless all voted" prop_checkVillagesTurnDoesNothingUnlessAllVoted,
 
-    testProperty "PROP: check werewolves turn advances to villagers" prop_checkWerewolvesTurnAdvancesToVillagers,
+    testProperty "PROP: check werewolves turn advances to villages" prop_checkWerewolvesTurnAdvancesToVillages,
     testProperty "PROP: check werewolves turn kills one player when consensus" prop_checkWerewolvesTurnKillsOnePlayerWhenConsensus,
     testProperty "PROP: check werewolves turn kills no one when conflicted" prop_checkWerewolvesTurnKillsNoOneWhenConflicted,
     testProperty "PROP: check werewolves turn resets votes" prop_checkWerewolvesTurnResetsVotes,
     testProperty "PROP: check werewolves turn does nothing unless all voted" prop_checkWerewolvesTurnDoesNothingUnlessAllVoted,
 
-    testProperty "PROP: check game over advances turn" prop_checkGameOverAdvancesTurn,
+    testProperty "PROP: check game over advances stage" prop_checkGameOverAdvancesStage,
     testProperty "PROP: check game over does nothing when at least two allegiances alive" prop_checkGameOverDoesNothingWhenAtLeastTwoAllegiancesAlive,
 
-    testProperty "PROP: start game starts with nightfall turn" prop_startGameStartsWithNightfallTurn,
+    testProperty "PROP: start game starts with sunset stage" prop_startGameStartsWithSunsetStage,
     testProperty "PROP: start game uses given players" prop_startGameUsesGivenPlayers,
     testProperty "PROP: start game errors unless unique player names" prop_startGameErrorsUnlessUniquePlayerNames,
     testProperty "PROP: start game errors when less than 7 players" prop_startGameErrorsWhenLessThan7Players,
@@ -109,7 +109,7 @@ allEngineTests = [
 
 allGameTests :: [TestTree]
 allGameTests = [
-    testProperty "PROP: new game starts with nightfall turn" prop_newGameStartsWithNightfallTurn,
+    testProperty "PROP: new game starts with sunset stage" prop_newGameStartsWithSunsetStage,
 
     testProperty "PROP: new game starts with sees empty" prop_newGameStartsWithSeesEmpty,
     testProperty "PROP: new game starts with votes empty" prop_newGameStartsWithVotesEmpty,